Move folder coach-profile-editor, create LevelList and QuestionList
