+++ src/app/models/coach-profile-editor/coach-profile-editor.component.scss
@@ -1,4 +1,4 @@
-@import 'src/theme/variables';
+@import '../../../theme/variables';
 
 .mat-elevation-z8 {
   margin: 10% 100px;
