+++ src/app/models/coach-profile-editor/coach-profile-editor.module.ts
@@ -4,13 +4,13 @@ import { MatIconModule } from '@angular/material/icon';
 import { MatOptionModule } from '@angular/material/core';
 import { MatSelectModule } from '@angular/material/select';
 import { MatCheckboxModule } from '@angular/material/checkbox';
-import { AngularMaterialCommonModule } from '../../angular-material-common.module';
+import { AngularMaterialCommonModule } from '../angular-material-common.module';
 import { CoachProfileEditorRoutingModule } from './coach-profile-editor-routing.module';
 import { EditorTableComponent } from './editor-table/editor-table.component';
 import { CoachProfileEditorComponent } from './coach-profile-editor.component';
 import { TopicAddingEditingDialogComponent } from './topic-adding-editing-dialog/topic-adding-editing-dialog.component';
 import { GrammarAddingEditingDialogComponent } from './grammar-adding-editing-dialog/grammar-adding-editing-dialog.component';
-import { ChooseLevelModule } from '../../../components/choose-level/choose-level.module';
+import { ChooseLevelModule } from '../../components/choose-level/choose-level.module';
 
 @NgModule({
   declarations: [
