+++ src/app/models/coach-profile-editor/editor-table/editor-table.component.scss
@@ -1,4 +1,4 @@
-@import '../../../../../theme/variables';
+@import '../../../../theme/variables';
 
 table {
   width: 100%;
